refactor(sdk): extract interpolation.py and system_spec.py from perf_database
Phase 1 of the SDK refactor. Decomposes two concerns out of perf_database.py
into focused modules so future operation migrations have stable primitives
to import:

- interpolation.py: 16 stateless functions for 1D/2D/3D interpolation,
  bilinear, grid extrapolation, top-k regime-aware piecewise interpolation
  (DSA + DeepSeek-V4 CSA), and the analytical estimate_mem_op helper used
  by ElementWise / Embedding.
- system_spec.py: SystemSpec(dict) wrapping the per-system YAML, with a
  promoted get_p2p_bandwidth() method (formerly PerfDatabase._get_p2p_bandwidth).
  Subclasses dict so existing system_spec["gpu"]["mem_bw"] access patterns
  remain unchanged.
- perf_database.py: methods become thin delegators (-635 lines). Internal
  call sites still go through self._foo so phase-2 op migrations are
  unaffected. Dead _extract_metric_data_3d removed.

Behavior is bit-identical to the pre-refactor code:
- estimate_mem_op consults gpu_spec lazily so SOL paths only need mem_bw and
  empirical paths still raise KeyError loudly on misconfigured specs.
- get_p2p_bandwidth uses direct dict indexing so missing topology keys raise
  KeyError, matching the original.

Validated against a 40-snapshot regression suite covering direct queries,
model construction, and end-to-end cli_estimate runs across LLAMA, Qwen3 MoE,
and DeepSeek V3/V3.2; plus the existing 564 unit tests under tests/unit/sdk.
